requirements.txt
disnake
PyDrive
PyYAML